♻️  userlist 추상화
