+++ src/pages/layouts/Searchuserhead.jsx
@@ -5,15 +5,11 @@ import { useHistory } from 'react-router-dom';
 
 export const Searchuserhead = ({ handleSearch }) => {
   let history = useHistory();
-  // const [search, searchInp] = useState();
 
   const back = () => {
     history.goBack();
   };
 
-  // const searchUser = (e) => {
-  //   console.log(e.target.value);
-  // };
   return (
     <Searchuserheads>
       <button onClick={back}>
